% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHero.R
\name{get_hero}
\alias{get_hero}
\title{Get a radom superhero name}
\usage{
get_hero()
}
\value{
A character string with the name of a superhero
}
\description{
Get a radom superhero name
}
\examples{
get_hero()
}
