% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superheroes}
\alias{superheroes}
\title{Vector of superhero names}
\format{
An object of class \code{character} of length 8.
}
\usage{
superheroes
}
\description{
A character vector containing names of superheroes.
}
\keyword{datasets}
